extract Station into an interface/ rename the enum to Stations instead of Station
